requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
stripe==7.12.0
passlib==1.7.4
sqlalchemy==2.0.25
PyJWT
mysql-connector-python==8.2.0
boto3==1.34.32
python-multipart==0.0.6
firebase-admin==6.4.0
